chore(docker): switch base image from Alpine to Debian
This change updates the Dockerfile for the socat-socks5h service to use the `debian:bookworm-slim` image instead of `alpine:3.19`. The package installation method was adjusted accordingly, replacing `apk` with `apt-get`, and includes removing the apt lists to reduce the image size. Additionally, renamed the entrypoint script from `entrypoint.sh` to `entrypoint.bash` to match convention or reflect a specific script change. This might lead to an increase in image size but ensures better compatibility with packages or libraries not available in Alpine.